# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Intelligence Platform",
    "description": "Cryptocurrency Market Intelligence & Quantitative Workstation v3.0. Factor Scanner, Robust Backtesting, Real-time Streaming, and Paper Trading. Non-custodial research platform. No live execution.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/scanner/rankings": {
      "get": {
        "tags": [
          "Live Scanner"
        ],
        "summary": "Get scanner rankings via single-query aggregation with pagination & filtering",
        "description": "Retrieve scanner rankings using a single unified query with window functions.\n\nGuarantees O(1) constant database queries regardless of universe size,\neliminating N+1 query performance degradation.",
        "operationId": "get_scanner_rankings_api_v1_scanner_rankings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by primary sector (e.g. L1, DeFi)",
              "title": "Sector"
            },
            "description": "Filter by primary sector (e.g. L1, DeFi)"
          },
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by asset class (e.g. CORE, ALTCOIN)",
              "title": "Asset Class"
            },
            "description": "Filter by asset class (e.g. CORE, ALTCOIN)"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum opportunity score filter",
              "title": "Min Score"
            },
            "description": "Minimum opportunity score filter"
          },
          {
            "name": "data_fresh_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Filter only assets with fresh market data",
              "default": false,
              "title": "Data Fresh Only"
            },
            "description": "Filter only assets with fresh market data"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field: opportunity_score, price, return_1d, return_7d, return_30d, symbol",
              "default": "opportunity_score",
              "title": "Sort By"
            },
            "description": "Sort field: opportunity_score, price, return_1d, return_7d, return_30d, symbol"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort direction"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScannerRankingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scanner/status": {
      "get": {
        "tags": [
          "Live Scanner"
        ],
        "summary": "Scanner operational status and data freshness summary",
        "description": "Return scanner data mode, freshness, and operational state.",
        "operationId": "get_scanner_status_api_v1_scanner_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Scanner Status Api V1 Scanner Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scanner/snapshots": {
      "get": {
        "tags": [
          "Live Scanner"
        ],
        "summary": "Historical scanner snapshot records with pagination & filtering",
        "description": "Return paginated historical scanner snapshot rows.",
        "operationId": "get_scanner_snapshots_api_v1_scanner_snapshots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Scanner Snapshots Api V1 Scanner Snapshots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scanner": {
      "get": {
        "tags": [
          "Scoring & Scanner"
        ],
        "summary": "Market Scanner Filter",
        "description": "Multi-parameter quantitative scanner.",
        "operationId": "scanner_api_v1_scanner_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssetClass"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by class",
              "title": "Asset Class"
            },
            "description": "Filter by class"
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by sector",
              "title": "Sector"
            },
            "description": "Filter by sector"
          },
          {
            "name": "min_opportunity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Minimum opportunity threshold",
              "default": 60.0,
              "title": "Min Opportunity"
            },
            "description": "Minimum opportunity threshold"
          },
          {
            "name": "max_risk_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter max risk",
              "title": "Max Risk Score"
            },
            "description": "Filter max risk"
          },
          {
            "name": "exclude_flags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Exclude assets with active risk flags",
              "default": false,
              "title": "Exclude Flags"
            },
            "description": "Exclude assets with active risk flags"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreSummaryResponse"
                  },
                  "title": "Response Scanner Api V1 Scanner Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetClass": {
        "type": "string",
        "enum": [
          "CORE",
          "LARGE_CAP_ALT",
          "ALTCOIN",
          "MID_CAP",
          "SMALL_CAP",
          "MEME",
          "UNKNOWN"
        ],
        "title": "AssetClass"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScannerRankingsResponse": {
        "properties": {
          "data_mode": {
            "type": "string",
            "title": "Data Mode"
          },
          "scanner_status": {
            "type": "string",
            "title": "Scanner Status"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "fresh_count": {
            "type": "integer",
            "title": "Fresh Count"
          },
          "stale_count": {
            "type": "integer",
            "title": "Stale Count"
          },
          "checked_at": {
            "type": "string",
            "title": "Checked At"
          },
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/ScannerRankingItem"
            },
            "type": "array",
            "title": "Rankings"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "scanner_status",
          "total_count",
          "limit",
          "offset",
          "fresh_count",
          "stale_count",
          "checked_at",
          "rankings"
        ],
        "title": "ScannerRankingsResponse",
        "description": "Paginated scanner rankings response."
      },
...
      "ScoreSummaryResponse": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "asset_class": {
            "type": "string",
            "title": "Asset Class"
          },
          "primary_sector": {
            "type": "string",
            "title": "Primary Sector"
          },
          "model_type": {
            "type": "string",
            "title": "Model Type"
          },
          "opportunity_score": {
            "type": "number",
            "title": "Opportunity Score"
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "trend_score": {
            "type": "number",
            "title": "Trend Score"
          },
          "momentum_score": {
            "type": "number",
            "title": "Momentum Score"
          },
          "relative_strength_score": {
            "type": "number",
            "title": "Relative Strength Score"
          },
          "liquidity_score": {
            "type": "number",
            "title": "Liquidity Score"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Flags"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "partial_data": {
            "type": "boolean",
            "title": "Partial Data",
            "default": false
          },
          "missing_inputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Inputs",
            "default": []
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "default": "v3.0.0"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "symbol",
          "name",
          "asset_class",
          "primary_sector",
          "model_type",
          "opportunity_score",
          "quality_score",
          "risk_score",
          "trend_score",
          "momentum_score",
          "relative_strength_score",
          "liquidity_score",
          "risk_flags",
          "time"
        ],
        "title": "ScoreSummaryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}